+++ src/common/js/index.js
@@ -15,7 +15,6 @@ import { TypedMessage } from '../classes/typed-message.class';
 
 import './global.js';
 import './constants.js';
-import './i18n.js';
 //
 import './configure.js';