+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Taxonomy change: the 'type' becomes a 'level' to be consistent with syslog(3)
     - Remove (unused) internationalization
     - iTypedMessageType() method is renamed iTypedMessageLevel()
+    - Remove MessagesSet class from the package
 
 ### 1.1.3