+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - iTypedMessageType() method is renamed iTypedMessageLevel()
     - Remove MessagesSet class from the package
     - Get rid of pwix:orderable-stack dependency
+    - Obsolete and remove the TypedMessage Blaze component
 
 ### 1.1.3